Server/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0

# API Documentation
flasgger==0.9.7.1

# Database dependencies
psycopg2-binary==2.9.9

# Testing dependencies
pytest==7.4.3
pytest-json-report==1.5.0

# Test reporting and analysis dependencies
pandas==2.1.3
Jinja2==3.1.2

flask-jwt-extended

werkzeug